replaced firebase with supabase

+++ lib/bloc/details_bloc/details_bloc.dart
@@ -1,20 +1,19 @@
 import 'dart:developer';
 
-import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:file_picker/file_picker.dart';
-import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:fluttertoast/fluttertoast.dart';
+import 'package:supabase_flutter/supabase_flutter.dart';
 import 'package:umair_liaqat/bloc/details_bloc/details_events.dart';
 import 'package:umair_liaqat/bloc/details_bloc/details_state.dart';
 import 'package:umair_liaqat/models/job_history.dart';
 import 'package:umair_liaqat/models/project_model.dart';
 import 'package:umair_liaqat/models/qualification_model.dart';
 import 'package:umair_liaqat/models/user_model.dart';
-import 'package:umair_liaqat/services/google_drive_service.dart';
 import 'package:umair_liaqat/services/media_service.dart';
 import 'package:umair_liaqat/utils/app_strings.dart';
+import 'package:umair_liaqat/utils/collections.dart';
 import 'package:umair_liaqat/utils/toast_utils.dart';
 import 'package:uuid/uuid.dart';
 
@@ -42,7 +41,10 @@ class DetailsBloc extends Bloc<DetailsEvents, DetailsState> {
     try {
       final value = await MediaService.selectFile(imageExtensions);
       if (value != null) {
-        String? uploadedProfileLink = await uploadPlatformFile(value);
+        String? uploadedProfileLink = await MediaService.uploadPlatformFile(
+          value,
+          isForProfile: true,
+        );
 
         emit(
           state.copyWith(
@@ -86,11 +88,16 @@ class DetailsBloc extends Bloc<DetailsEvents, DetailsState> {
       UserDataUpdateEvent event, Emitter<DetailsState> emit) async {
     try {
       ToastUtils.showLoader(event.context);
+      final client = Supabase.instance.client;
+      final userEmail = client.auth.currentUser?.email;
 
-      final User? user = FirebaseAuth.instance.currentUser;
+      if (userEmail == null) {
+        throw Exception("User not authenticated");
+      }
       final UserModel userModel = UserModel(
         name: event.name,
         description: event.description,
+        email: userEmail,
         headline1: event.headline1,
         headline2: event.headline2,
         profilePicture: event.profilePicture,
@@ -98,15 +105,36 @@ class DetailsBloc extends Bloc<DetailsEvents, DetailsState> {
         linkedIn: event.linkedIn,
         phoneNumber: event.phoneNumber,
       );
-      await FirebaseFirestore.instance
-          .collection(DatabaseCollections.user)
-          .doc(user!.uid)
-          .set(
-            userModel.toMap(),
-          );
+      final userResponse = await Supabase.instance.client
+          .from(Collections.users)
+          .select()
+          .limit(1)
+          .maybeSingle();
+      if (userResponse == null) {
+        final updateResponse = await client
+            .from('users')
+            .insert(userModel.toMapSimpleUser())
+            .eq('email', userEmail);
+        if (updateResponse != null) {
+          debugPrint(updateResponse);
+        }
+      } else {
+        final updateResponse = await client
+            .from('users')
+            .update(userModel.toMapSimpleUser())
+            .eq('email', userEmail);
+        if (updateResponse != null) {
+          debugPrint(updateResponse);
+        }
+      }
+
       Navigator.of(event.context).pop();
 
-      Fluttertoast.showToast(msg: Strings.valueAdded(Strings.workHistory));
+      Fluttertoast.showToast(
+        msg: Strings.valueUpdated(
+          Strings.userDetails,
+        ),
+      );
     } catch (e) {
       Navigator.of(event.context).pop();
 
@@ -118,57 +146,65 @@ class DetailsBloc extends Bloc<DetailsEvents, DetailsState> {
 
   Future<void> _uploadWorkHistory(
       UploadWorkHistory event, Emitter<DetailsState> emit) async {
+    final userId = Supabase.instance.client.auth.currentUser?.id;
+    if (userId == null) {
+      Fluttertoast.showToast(msg: "User not authenticated");
+      return;
+    }
+
     String id = _uuid.v4();
     try {
       ToastUtils.showLoader(event.context);
+
       JobHistory jobHistory = JobHistory(
         id: id,
         fromDate: event.fromDate,
+        toDate: event.toDate,
         jobDescription: event.description,
         organization: event.organization,
         position: event.jobPosition,
         sortIndex: event.sortIndex,
-        toDate: event.toDate,
+        userId: userId, // Make sure your model has this
       );
-      await FirebaseFirestore.instance
-          .collection(DatabaseCollections.jobHistory)
-          .doc(id)
-          .set(
-            jobHistory.toMap(),
-          );
-      Navigator.of(event.context).pop();
 
+      await Supabase.instance.client
+          .from(Collections.jobHistory)
+          .insert(jobHistory.toMap());
+
+      Navigator.of(event.context).pop();
       Fluttertoast.showToast(msg: Strings.valueAdded(Strings.workHistory));
     } catch (e) {
       Navigator.of(event.context).pop();
-
       Fluttertoast.showToast(msg: e.toString());
-
       log("Error while uploading work history: $e");
     }
   }
 
   Future<void> _updateWorkHistory(
       UpdateWorkHistory event, Emitter<DetailsState> emit) async {
+    final userId = Supabase.instance.client.auth.currentUser?.id;
+    if (userId == null) {
+      Fluttertoast.showToast(msg: "User not authenticated");
+      return;
+    }
+
     try {
       ToastUtils.showLoader(event.context);
 
-      await FirebaseFirestore.instance
-          .collection(DatabaseCollections.jobHistory)
-          .doc(event.jobHistory.id)
-          .update(
-            event.jobHistory.toMap(),
-          );
+      await Supabase.instance.client
+          .from(Collections.jobHistory)
+          .update(event.jobHistory.toMap())
+          .eq('id', event.jobHistory.id!)
+          .eq('userId', userId); // ownership check
+
       Navigator.of(event.context).pop();
       Navigator.of(event.context).pop();
 
       Fluttertoast.showToast(msg: Strings.valueUpdated(Strings.workHistory));
     } catch (e) {
       Navigator.of(event.context).pop();
-
       Fluttertoast.showToast(msg: e.toString());
-
-      log("Error while uploading work history: $e");
+      log("Error while updating work history: $e");
     }
   }
 
@@ -206,74 +242,90 @@ class DetailsBloc extends Bloc<DetailsEvents, DetailsState> {
 
   Future<void> _uploadQualification(
       UploadQualification event, Emitter<DetailsState> emit) async {
-    String id = _uuid.v4();
+    final userId = Supabase.instance.client.auth.currentUser?.id;
+    if (userId == null) {
+      Fluttertoast.showToast(msg: "User not authenticated");
+      return;
+    }
+
+    final id = _uuid.v4();
     try {
       ToastUtils.showLoader(event.context);
 
-      QualificationModel qualificationModel = QualificationModel(
+      final qualificationModel = QualificationModel(
         id: id,
         completionYear: event.completionYear,
         degreeName: event.degreeName,
         instituteName: event.institute,
         sortingIndex: event.sortIndex,
       );
-      await FirebaseFirestore.instance
-          .collection(DatabaseCollections.qualifications)
-          .doc(id)
-          .set(
-            qualificationModel.toMap(),
-          );
-      Navigator.of(event.context).pop();
 
+      final data = qualificationModel.toMap()..addAll({'userId': userId});
+
+      await Supabase.instance.client
+          .from(Collections.qualifications)
+          .insert(data);
+
+      Navigator.of(event.context).pop();
       Fluttertoast.showToast(msg: Strings.valueAdded(Strings.qualification));
     } catch (e) {
       Navigator.of(event.context).pop();
-
       Fluttertoast.showToast(msg: e.toString());
-
       log("Error while uploading qualification: $e");
     }
   }
 
   Future<void> _updateQualification(
       UpdateQualification event, Emitter<DetailsState> emit) async {
+    final userId = Supabase.instance.client.auth.currentUser?.id;
+    if (userId == null) {
+      Fluttertoast.showToast(msg: "User not authenticated");
+      return;
+    }
+
     try {
       ToastUtils.showLoader(event.context);
 
-      await FirebaseFirestore.instance
-          .collection(DatabaseCollections.qualifications)
-          .doc(event.qualificationModel.id)
-          .update(
-            event.qualificationModel.toMap(),
-          );
-      Navigator.of(event.context).pop();
-      Navigator.of(event.context).pop();
+      event.qualificationModel =
+          event.qualificationModel.copyWith(userId: userId);
 
+      await Supabase.instance.client
+          .from(Collections.qualifications)
+          .update(event.qualificationModel.toMap())
+          .eq('id', event.qualificationModel.id!)
+          .eq('userId', userId);
+
+      Navigator.of(event.context).pop();
+      Navigator.of(event.context).pop();
       Fluttertoast.showToast(msg: Strings.valueUpdated(Strings.qualification));
     } catch (e) {
       Navigator.of(event.context).pop();
-
       Fluttertoast.showToast(msg: e.toString());
-
-      log("Error while uploading qualification: $e");
+      log("Error while updating qualification: $e");
     }
   }
 
   Future<void> _deleteQualification(
       DeleteQualification event, Emitter<DetailsState> emit) async {
+    final userId = Supabase.instance.client.auth.currentUser?.id;
+    if (userId == null) {
+      Fluttertoast.showToast(msg: "User not authenticated");
+      return;
+    }
+
     try {
       ToastUtils.showLoader(event.context);
 
-      await FirebaseFirestore.instance
-          .collection(DatabaseCollections.qualifications)
-          .doc(event.id)
-          .delete();
-      Navigator.of(event.context).pop();
+      await Supabase.instance.client
+          .from(Collections.qualifications)
+          .delete()
+          .eq('id', event.id)
+          .eq('userId', userId);
 
+      Navigator.of(event.context).pop();
       Fluttertoast.showToast(msg: Strings.valueDeleted(Strings.qualification));
     } catch (e) {
       Navigator.of(event.context).pop();
-
       Fluttertoast.showToast(msg: e.toString());
       log("Error while deleting qualification: $e");
     }
@@ -281,36 +333,44 @@ class DetailsBloc extends Bloc<DetailsEvents, DetailsState> {
 
   Future<void> _deleteJobHistory(
       DeleteWorkHistory event, Emitter<DetailsState> emit) async {
+    final userId = Supabase.instance.client.auth.currentUser?.id;
+    if (userId == null) {
+      Fluttertoast.showToast(msg: "User not authenticated");
+      return;
+    }
+
     try {
       ToastUtils.showLoader(event.context);
 
-      await FirebaseFirestore.instance
-          .collection(DatabaseCollections.jobHistory)
-          .doc(event.id)
-          .delete();
+      await Supabase.instance.client
+          .from(Collections.jobHistory)
+          .delete()
+          .eq('id', event.id)
+          .eq('userId', userId); // prevent deleting others' data
+
       Navigator.of(event.context).pop();
       Fluttertoast.showToast(msg: Strings.valueDeleted(Strings.jobHistory));
     } catch (e) {
       Navigator.of(event.context).pop();
       Fluttertoast.showToast(msg: e.toString());
       log("Error while deleting job history: $e");
-      Fluttertoast.showToast(msg: e.toString());
     }
   }
 
   Future<void> _deleteProject(
-      DeleteProject event, Emitter<DetailsState> emit) async {
+    DeleteProject event,
+    Emitter<DetailsState> emit,
+  ) async {
     try {
       ToastUtils.showLoader(event.context);
 
-      await FirebaseFirestore.instance
-          .collection(DatabaseCollections.projects)
-          .doc(event.id)
-          .delete();
+      await Supabase.instance.client
+          .from('projects')
+          .delete()
+          .eq('id', event.id);
+
       Navigator.of(event.context).pop();
-      Fluttertoast.showToast(
-        msg: Strings.valueDeleted(Strings.project),
-      );
+      Fluttertoast.showToast(msg: Strings.valueDeleted(Strings.project));
     } catch (e) {
       Navigator.of(event.context).pop();
       Fluttertoast.showToast(msg: e.toString());
@@ -319,86 +379,84 @@ class DetailsBloc extends Bloc<DetailsEvents, DetailsState> {
   }
 
   Future<void> _uploadProject(
-      UploadProjectEvent event, Emitter<DetailsState> emit) async {
-    String id = _uuid.v4();
+    UploadProjectEvent event,
+    Emitter<DetailsState> emit,
+  ) async {
+    final id = _uuid.v4();
     try {
+      final userId = Supabase.instance.client.auth.currentUser?.id;
+      if (userId == null) {
+        Fluttertoast.showToast(msg: "User not authenticated");
+        return;
+      }
       ToastUtils.showLoader(event.context);
+
       List<String> linksList = [];
+
       if (event.projectModel.files != null &&
           event.projectModel.files!.isNotEmpty) {
-        for (int i = 0; i < (event.projectModel.files?.length ?? 0); i++) {
-          PlatformFile file = event.projectModel.files![i];
-          String? link = await uploadPlatformFile(file);
+        for (PlatformFile file in event.projectModel.files!) {
+          String? link = await MediaService.uploadPlatformFile(file);
           if (link != null && link.isNotEmpty) {
             linksList.add(link);
           }
         }
       }
+
       ProjectModel projectModel = event.projectModel.copyWith(
         id: id,
         filesLinks: linksList,
+        userId: userId,
       );
 
-      await FirebaseFirestore.instance
-          .collection(DatabaseCollections.projects)
-          .doc(id)
-          .set(
-            projectModel.toMap(),
-          );
+      await Supabase.instance.client
+          .from('projects')
+          .insert(projectModel.toMap());
+
       Navigator.of(event.context).pop();
-      Fluttertoast.showToast(
-        msg: Strings.valueAdded(
-          Strings.project,
-        ),
-      );
+      Fluttertoast.showToast(msg: Strings.valueAdded(Strings.project));
     } catch (e) {
       Navigator.of(event.context).pop();
-
       Fluttertoast.showToast(msg: e.toString());
-
       log("Error while uploading project: $e");
     }
   }
 
   Future<void> _updateProject(
-      UpdateProjectEvent event, Emitter<DetailsState> emit) async {
+    UpdateProjectEvent event,
+    Emitter<DetailsState> emit,
+  ) async {
     try {
       ToastUtils.showLoader(event.context);
-      List<String> linksList = [];
+      List<String> newLinks = [];
+
       if (event.projectModel.files != null &&
           event.projectModel.files!.isNotEmpty) {
-        for (int i = 0; i < (event.projectModel.files?.length ?? 0); i++) {
-          PlatformFile file = event.projectModel.files![i];
-          String? link = await uploadPlatformFile(file);
+        for (PlatformFile file in event.projectModel.files!) {
+          String? link = await MediaService.uploadPlatformFile(file);
           if (link != null && link.isNotEmpty) {
-            linksList.add(link);
+            newLinks.add(link);
           }
         }
       }
-      ProjectModel projectModel = event.projectModel.copyWith(
+
+      ProjectModel updatedProject = event.projectModel.copyWith(
         filesLinks: [
-          ...event.projectModel.filesLinks ?? [],
-          ...linksList,
+          ...?event.projectModel.filesLinks,
+          ...newLinks,
         ],
       );
 
-      await FirebaseFirestore.instance
-          .collection(DatabaseCollections.projects)
-          .doc(event.projectModel.id)
-          .update(
-            projectModel.toMap(),
-          );
+      await Supabase.instance.client
+          .from('projects')
+          .update(updatedProject.toMap())
+          .eq('id', updatedProject.id!);
+
       Navigator.of(event.context).pop();
-      Fluttertoast.showToast(
-        msg: Strings.valueUpdated(
-          Strings.project,
-        ),
-      );
+      Fluttertoast.showToast(msg: Strings.valueUpdated(Strings.project));
     } catch (e) {
       Navigator.of(event.context).pop();
-
       Fluttertoast.showToast(msg: e.toString());
-
       log("Error while updating project: $e");
     }
   }
